Panic if decrement operator detected in C source code
With the introduction of the negation operator (represented by a single
hyphen in C source code), the hyphen character is now an accepted
character in the input C source code.

However, the decrement operator (represented by two consecutive hyphens
in C source code) consists of the now accepted hyphen character, despite
the decrement operator not yet being supported.

Being able to detect the decrement operator is necessary, in order to be
able to distinguish between:
- a C program containing the decrement operator (which is not yet
  supported)
- a C program containing the negation operator (which is having support
  added in)

in order to be able to reject the former, but accept the latter.