Parse C return of an expression containing two unary operators into IR instruction AST nodes
Note that, due to tests running in parallel, the `static mut
IDENTIFIER_COUNT` is shared across all threads running the tests.

This means that the two tests involving temporary variable IR AST nodes
both increment the value bound to `IDENTIFIER_COUNT`, causing the
asserted temporary variable names to be not as expected (more
specifically, the integer is always one more than expected in one of the
tests in every run), and thus one of these two tests is always failing
currently.