Rename current C expression parsing function
Some reshuffling of the C parsing functions is necessary to begin
parsing binary operators, one of which involves introducing a
distinction between a language construct called a "factor" vs. a
language construct called an "expression".

The language construct that has up until now been called an "expression"
is now being renamed to "factor", and a new language construct with the
name "expression" will be introduced shortly.