Fix incorrect construction of C statement
After the modifications to the C grammar to include "factor" in addition
to "expression" (when precendence climbing was introduced to parse
binary operator application), the production rules of some language
constructs changed a bit.

In particular, the production rules in the layers around statements and
expressions were in flux, and it looks like the production rule of a C
statement changed was incorrectly left to be using a "factor" rather
than an "expression".